+++ src/Repository/UserRepository.php
@@ -124,13 +124,6 @@ class UserRepository extends AbstractRepository
                 ->setParameter('lineItemSlugs', $criteria->getLineItemSlugCriterion());
         }
 
-        if ($criteria->hasEuclideanDivisionCriterion()) {
-            $queryBuilder
-                ->andWhere('MOD(u.id, :modulo) = :remainder')
-                ->setParameter('modulo', $criteria->getEuclideanDivisionCriterion()->getModulo())
-                ->setParameter('remainder', $criteria->getEuclideanDivisionCriterion()->getRemainder());
-        }
-
         $userIds = [];
         $usernameCollection = new UsernameCollection();
         $result = $queryBuilder->getQuery()->getResult();
@@ -169,13 +162,6 @@ class UserRepository extends AbstractRepository
                 ->leftJoin('a.lineItem', 'l')
                 ->andWhere('l.slug IN (:lineItemSlugs)')
                 ->setParameter('lineItemSlugs', $criteria->getLineItemSlugCriterion());
-        }
-
-        if ($criteria->hasEuclideanDivisionCriterion()) {
-            $queryBuilder
-                ->where('MOD(u.id, :modulo) = :remainder')
-                ->setParameter('modulo', $criteria->getEuclideanDivisionCriterion()->getModulo())
-                ->setParameter('remainder', $criteria->getEuclideanDivisionCriterion()->getRemainder());
         }
 
         $result = $queryBuilder